+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Fixed:
 - Added pollNow() for pmac GPIOs (requires pmac 2-1 or later)
 - Enforce camelCaseFields and Alphanumeric + underscore + dash Part names
 - Don't load child runnable block designs at init
+- Error message when a Field is not writeable now shows current state
 
 `3-0a3`_ - 2018-07-25
 ---------------------